feat(web): 更新发布飞机页面和预览组件以支持视频封面
- 在PublishAircraftLivePreview组件中引入coverMode以处理视频封面逻辑，优化媒体插槽构建。
- 更新PublishAircraftPage组件以支持视频封面选择，添加视频帧截取功能。
- 改进用户交互体验，确保在未选择封面时提供清晰的提示信息。
- 更新相关样式以增强视觉一致性和响应性。

+++ packages/schemas/src/aircraft-submissions.ts
@@ -37,10 +37,10 @@ export const aircraftSubmissionStatusSchema = z.enum([
 ]);
 
 export const aircraftSubmissionParametersSchema = z.object({
-  maxFlightTimeMinutes: z.number().nonnegative().nullable(),
-  maxRangeKilometers: z.number().nonnegative().nullable(),
-  maxSpeedKph: z.number().nonnegative().nullable(),
-  takeoffWeightGrams: z.number().nonnegative().nullable()
+  maxFlightTimeMinutes: z.number().int().nonnegative().nullable(),
+  maxRangeKilometers: z.number().int().nonnegative().nullable(),
+  maxSpeedKph: z.number().int().nonnegative().nullable(),
+  takeoffWeightGrams: z.number().int().nonnegative().nullable()
 });
 
 export const aircraftSubmissionVideoAssetSchema = z.object({
@@ -103,10 +103,10 @@ export const createAircraftSubmissionInputSchema = z.object({
   videoFileId: z.string().trim().min(1).nullable(),
   priceMin: priceValueSchema,
   priceMax: priceValueSchema,
-  maxFlightTimeMinutes: z.number().nonnegative().nullable(),
-  maxRangeKilometers: z.number().nonnegative().nullable(),
-  maxSpeedKph: z.number().nonnegative().nullable(),
-  takeoffWeightGrams: z.number().nonnegative().nullable()
+  maxFlightTimeMinutes: z.number().int().nonnegative().nullable(),
+  maxRangeKilometers: z.number().int().nonnegative().nullable(),
+  maxSpeedKph: z.number().int().nonnegative().nullable(),
+  takeoffWeightGrams: z.number().int().nonnegative().nullable()
 }).superRefine(validatePriceRange);
 
 export const updateAircraftSubmissionStatusInputSchema = z.object({
